restructuring the application
