+++ go.mod
@@ -3,15 +3,13 @@ module github.com/yannickalex07/dmon
 go 1.18
 
 require (
+	github.com/jellydator/ttlcache/v3 v3.0.0
 	github.com/sirupsen/logrus v1.9.0
 	github.com/slack-go/slack v0.11.2
 	google.golang.org/api v0.92.0
 )
 
-require (
-	github.com/gorilla/websocket v1.4.2 // indirect
-	github.com/jellydator/ttlcache/v3 v3.0.0 // indirect
-)
+require github.com/gorilla/websocket v1.4.2 // indirect
 
 require (
 	cloud.google.com/go/compute v1.7.0 // indirect
