+++ pkg/interfaces/api.go
@@ -1,6 +1,6 @@
 package interfaces
 
-import "github.com/yannickalex07/dmon/models"
+import "github.com/yannickalex07/dmon/pkg/models"
 
 type API interface {
 	Jobs(project string, location string) ([]models.Job, error)
