+++ pkg/models/entry.go
